relay-server/requirement.txt
fastapi==0.95.2
uvicorn==0.22.0
mysql-connector-python==8.0.33
pydantic==1.10.8
python-dotenv==1.0.0
openai==1.0.0

